Add files via upload
Reviewers' comments on the version of the paper submitted to IVAPP2 Conference 2022 with no possible rebuttal...
But My rebuttal:
Thank you for your reply and thanks to the reviewers for the time they took to read the paper and write their report. From this report, it appears that : 
- They were not able to use the tool and reproduce the results; 
- The tool does not display the results in the expert view and the novice view as advertised; 
- The difference with the previous paper was not discussed; 
- There is too much mathematics; 
- There is a need to conduct an evaluation study with a novice; 
In fact, apart from the last point, we find the other points justifiable and some are due to misunderstanding. 
For the first two points, for example, we have integrated the usage example in GitHub, we don't know if the reviewers read it and if they looked at the animatedGif we made; that serves as a usage guide before going to the live version. As for the difference with the previous paper, yes we could maybe have said something about it, but it would have been just to say that it's not the same problem we want to solve. The first paper aims to visualize code written with vector instructions while this paper aims to visualize vector instructions. Finally, we sincerely don't think that there is too much mathematics, in fact, the mathematics here is used to validate the data model used, we don't see how it could be otherwise! 
As for the last point precisely, one of the authors is a novice in the field of vector instructions and the other is an expert in this field. 
In the absence of the ability to rebut reviewer comments through the article submission interface, please consider this as such a rebuttal. 
Best regards.